Stores refactored
Frontend stores naming refactored
