+++ FrontEnd/src/app/company/store/company.store.state.ts
